Sepatate lib.rs to crab.rs + darwinia.rs
